refactor(recordings): decompose Sidebar resize and status-notification logic into focused hooks (FE-012)
